Added Indestructible characters in a few places (and at the end)
Red enemies are indestructible.  Added an additional menu option to jump
to an alternate final level where the indestructible character will pass
through you to make it easier to complete to demonstrate the end game
menu.
Also, I packaged the game for windows and I guess it left a bunch of
entries in the project folder :P